+++ src/arch/armv8/gic.c
@@ -27,7 +27,7 @@ void gicd_init()
     size_t int_num = gic_num_irqs();
 
     /* Bring distributor to known state */
-    for (int i = GIC_NUM_PRIVINT_REGS; i < GIC_NUM_INT_REGS(int_num); i++) {
+    for (size_t i = GIC_NUM_PRIVINT_REGS; i < GIC_NUM_INT_REGS(int_num); i++) {
         /**
          * Make sure all interrupts are not enabled, non pending,
          * non active.
@@ -38,14 +38,14 @@ void gicd_init()
     }
 
     /* All interrupts have lowest priority possible by default */
-    for (int i = GIC_NUM_PRIO_REGS(GIC_CPU_PRIV);
+    for (size_t i = GIC_NUM_PRIO_REGS(GIC_CPU_PRIV);
          i < GIC_NUM_PRIO_REGS(int_num); i++) {
         gicd.IPRIORITYR[i] = -1;
     }
 
     if (GIC_VERSION == GICV2) {
         /* No CPU targets for any interrupt by default */
-        for (int i = GIC_NUM_TARGET_REGS(GIC_CPU_PRIV);
+        for (size_t i = GIC_NUM_TARGET_REGS(GIC_CPU_PRIV);
              i < GIC_NUM_TARGET_REGS(int_num); i++) {
             gicd.ITARGETSR[i] = 0;
         }
@@ -54,7 +54,7 @@ void gicd_init()
         gicd.CTLR |= GICD_CTLR_EN_BIT;
 
     } else {
-        for (int i = GIC_CPU_PRIV; i < GIC_MAX_INTERUPTS; i++) {
+        for (size_t i = GIC_CPU_PRIV; i < GIC_MAX_INTERUPTS; i++) {
             gicd.IROUTER[i] = GICD_IROUTER_INV;
         }
